UI: Extracted TimeDecoratorDelegate from feature_splitting_window.py
